+++ src/main/java/ProgressoApp/controllers/ProjectController.java
@@ -2,10 +2,8 @@ package ProgressoApp.controllers;
 
 import ProgressoApp.model.Project;
 import ProgressoApp.service.ProjectService;
-import jakarta.validation.Valid;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
