+++ src/main/java/ProgressoApp/controllers/TaskController.java
@@ -3,7 +3,6 @@ package ProgressoApp.controllers;
 import ProgressoApp.dto.request.TaskRequestDTO;
 import ProgressoApp.model.Task;
 import ProgressoApp.service.TaskService;
-import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
