--- src/main/java/ProgressoApp/service/impl/TaskServiceImpl.java
@@ -1,55 +1,0 @@
-package ProgressoApp.service.impl;
-
-import ProgressoApp.dto.request.TaskRequestDTO;
-import ProgressoApp.model.Project;
-import ProgressoApp.model.Task;
-import ProgressoApp.model.TaskStatus;
-import ProgressoApp.repository.ProjectRepository;
-import ProgressoApp.repository.TaskRepository;
-import ProgressoApp.service.TaskService;
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-
-@Service
-public class TaskServiceImpl implements TaskService {
-
-  private final TaskRepository taskRepository;
-  private final ProjectRepository projectRepository;
-
-  @Autowired
-  public TaskServiceImpl(TaskRepository taskRepository, ProjectRepository projectRepository) {
-    this.taskRepository = taskRepository;
-    this.projectRepository = projectRepository;
-  }
-
-
-  public Optional<Task> getTaskById(Long id) {
-    return taskRepository.findById(id);
-  }
-
-
-  public Task createTask(TaskRequestDTO dto) {
-    Task task = new Task();
-
-    return taskRepository.save(task);
-  }
-
-  public Task updateTask(Long id, Task updatedTask) {
-    return taskRepository.findById(id)
-        .map(task -> {
-          task.setName(updatedTask.getName());
-          task.setDescription(updatedTask.getDescription());
-          task.setTaskOrder(updatedTask.getTaskOrder());
-          task.setTaskStatus(updatedTask.getTaskStatus());
-          return taskRepository.save(task);
-        }).orElseThrow(() -> new RuntimeException("Task not found"));
-  }
-
-  public void deleteTask(Long id) {
-    taskRepository.deleteById(id);
-  }
-
-}
